FE/Qt: bugref:10322: Runtime UI: Cleanup for arguments of mouse/keyboard wrappers.
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@98601 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Frontends/VirtualBox/src/runtime/UIMouseHandler.cpp
@@ -1261,7 +1261,7 @@ bool UIMouseHandler::multiTouchEvent(QTouchEvent *pTouchEvent, ulong uScreenId)
     uimachine()->putEventMultiTouch(pTouchEvent->touchPoints().size(),
                                     contacts,
                                     fTouchScreen,
-                                    (ULONG)RTTimeMilliTS());
+                                    RTTimeMilliTS());
 
     /* Eat by default? */
     return true;
